Add some identify command content, remove quark link
Can't say we're using quark anymore, because quark doesn't deal well with urls
in the format example.com/?blahblah. It tries to search for the file ?blahblah
and returns a 404. No good if facebook wants to link to your site, because they
add one of those to make sure they have their claws in everything.